digraph: refactor
