+++ internal/digraph/constants.go
@@ -1,0 +1,13 @@
+// Copyright (C) 2025 Yota Hamada
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package digraph
+
+// Special environment variables.
+const (
+	EnvKeyLogPath          = "DAG_EXECUTION_LOG_PATH"
+	EnvKeySchedulerLogPath = "DAG_SCHEDULER_LOG_PATH"
+	EnvKeyRequestID        = "DAG_REQUEST_ID"
+	EnvKeyDAGName          = "DAG_NAME"
+	EnvKeyDAGStepName      = "DAG_STEP_NAME"
+)
