+++ internal/digraph/template.go
@@ -1,3 +1,6 @@
+// Copyright (C) 2025 Yota Hamada
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 package digraph
 
 import (
